Add Ubuntu based Ruby 2.4 image
For instances where we want test parity with our Ubuntu based Ruby
deployments we'd like an alternative to the existing alpine image.

This provides that. It installs a few common packages, including the
appropriate system libraries for nokogiri. en_AU.UTF-8 is configured as
the default language. The postgres apt repository is configured but we
do not install anything from it.