Build Ruby 2.6 base image
Support for Ruby 2.4 has now ended, which means:

> Bug and security fixes from more recent Ruby versions will no longer
> be backported to 2.4, and no further patch release of 2.4 will be
> released.

https://www.ruby-lang.org/en/news/2020/04/05/support-of-ruby-2-4-has-ended/

That's a problem for us, so we want to upgrade to a supported version of
Ruby.

For now we want to keep using our Ubuntu 16.04 base, so we'll keep using
Ruby from the brightbox PPA. This PPA provides a Ruby 2.6 package but
not one for 2.7 yet, so we'll start by upgrading to Ruby 2.6.

https://launchpad.net/~brightbox/+archive/ubuntu/ruby-ng
https://www.ruby-lang.org/en/news/2017/12/25/ruby-2-5-0-released/
https://www.ruby-lang.org/en/news/2018/12/25/ruby-2-6-0-released/