Build node 12.17.0 image
Our current image provides nodejs 12.14.0. We'd like to upgrade to take
advantage of some enhancements to memory management, which may help with
what may be a leak in one application.

This also starts tagging the image with the node version rather than the
alpine one; this makes it a bit easier to see at a glance what's being
provided and should be considered an immutable tag.

See:
* https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V12.md
* https://github.com/nodejs/node/pull/27345