Build Ruby 2.7 base image
Ruby 3.0 was released on December 25, 2020. We're currently using Ruby
2.6, which is two years old now. While we don't have a pressing need to
upgrade to 2.7 like we did when moving away from 2.4, it's generally
good for us to stay on top of these things.

The Ruby 3.0 release blog post mentions that:

> In principle, code that prints a warning on Ruby 2.7 won’t work

Which is also a good motivation for us to make the upgrade gradually -
we probably won't get to 3.0 anytime soon, but we can make that future
upgrade smaller by getting on 2.7 first.

See:
* https://www.ruby-lang.org/en/news/2019/12/25/ruby-2-7-0-released/
* https://www.ruby-lang.org/en/news/2020/12/25/ruby-3-0-0-released/